Move graphql query transformation to separate class
